Updated UI
So, I really do not know how to add all the individual files to the GitHub so I just added the .zip :( sorry yall!
What is on this is a welcome page and then a page with game selection, if I have left anything out or if it is showing weird on your emulator let me know and I'll fix it. 
Also, if the buttons and images look like they are all up in the corner when you are on design mode for the activity_select_game.xml, I do not know why it's doing that but still compiling and showing correctly in the emulator. The same goes for if it tells you there are errors and unknown id's for layout_below and things of that nature.
I'm going to work on the "How to play" buttons next